refactor: move DeleteGroupDialogContent to the components root directory
