--- src/features/groups/components/index.ts
@@ -1,2 +1,0 @@
-export { CreateGroupButton } from './CreateGroupButton';
-export { GroupCardList } from './GroupCardList';
